Tag rest api + test, some more fixes, removed success report for rest responses

+++ modules/api/controllers/actions/TagActions.php
@@ -5,28 +5,7 @@ namespace app\modules\api\controllers;
 use app\models\Image;
 use app\models\Tag;
 
-class BindTagAction extends \yii\rest\Action
-{
-    public function run($id)
-    {
-        /** @var Image $image */
-        $image = Image::findOne($id);
-        if (!$image)
-        {
-            return \Yii::$app->response->setStatusCode(404);
-        }
-        /** @var Tag $tagName */
-        $tagName = \Yii::$app->request->bodyParams["name"];
-        $tag = Tag::findOne(["name" => $tagName]);
-        if (!$tag)
-        {
-            return \Yii::$app->response->setStatusCode(400);
-        }
-        $image->link('tags', $tag);
-    }
-}
-
-class UnbindTagAction extends \yii\rest\Action
+class ToggleTagAction extends \yii\rest\Action
 {
     public function run($id, $tagName)
     {
@@ -34,14 +13,24 @@ class UnbindTagAction extends \yii\rest\Action
         $image = Image::findOne($id);
         if (!$image)
         {
-            return \Yii::$app->response->setStatusCode(404);
+            return \Yii::$app->response->setStatusCode(404, 'Image not found');
         }
-        /** @var Tag $tagName */
+        /** @var Tag $tag */
         $tag = Tag::findOne(["name" => $tagName]);
         if (!$tag)
         {
-            return \Yii::$app->response->setStatusCode(400);
+            return \Yii::$app->response->setStatusCode(404, 'Tag not found');
         }
-        $image->unlink('tags', $tag, true);
+
+        $method = \Yii::$app->request->method;
+
+        if ($method === 'POST' && !$image->hasTag($tag->name))
+        {
+            $image->link('tags', $tag);
+        }
+        else if ($method === 'DELETE')
+        {
+            $image->unlink('tags', $tag, true);
+        }
     }
 }
